Linux, Android.
-Added a linux console app project.
-Added an android dynamic shared library project.
-Changed some datatypes in library to make it work in all targeted platforms (currently cannot build the android project due to cmath library).

+++ HephAudio/AudioFormats.h
@@ -18,7 +18,7 @@ namespace HephAudio
 				uint32_t cursor = 0;
 				while (true)
 				{
-					const uint32_t pos = extension.find(' ', cursor);
+					const size_t pos = extension.find(' ', cursor);
 					splitExtensions.push_back(extension.substr(cursor, pos - cursor));
 					cursor = pos + 1;
 					if (pos == std::wstring::npos)
